moved wizard photos to module in imgs, created update logic for 'Update' event on client
