+++ client/www/assets/imgs/wizardPhotos.js
@@ -1,0 +1,12 @@
+var wizardPhotos = [
+  'walking_wizard.gif',
+  'simpsons_wizard.jpg',
+  'wizard_by_adam_brown.jpg',
+  'cartman_wizard.png',
+  'character_wizard.png',
+  'DC_wizard.png',
+  'eggplant_wizard_uprising.png',
+  'evil_wizard.png',
+  'merlin_the_wizard.png',
+]
+module.exports = wizardPhotos;
